% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modfit.r
\name{model_fitting}
\alias{model_fitting}
\title{Add Two Numbers}
\usage{
model_fitting(mod, dat)
}
\arguments{
\item{x}{A numeric, the first number to add.}

\item{y}{A numeric, the second number to add.}
}
\value{
A numeric, the sum of x and y.
}
\description{
This function takes two numbers and returns their sum.
}
