% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotts.R
\name{plot.ts}
\alias{plot.ts}
\title{Add Two Numbers}
\usage{
\method{plot}{ts}(dat)
}
\arguments{
\item{x}{A numeric, the first number to add.}

\item{y}{A numeric, the second number to add.}
}
\value{
A numeric, the sum of x and y.
}
\description{
This function takes two numbers and returns their sum.
}
